MR-209: problematic advance
