+++ src/Components/Header/header.module.css
@@ -24,6 +24,14 @@
   margin-right: 1rem;
 }
 
+.logOut {
+  width: 24px;
+  height: 24px;
+  margin-left: 5px;
+  margin-right: 1rem;
+background-color: red;
+}
+
 .socialIcon {
   margin-left: 5px;
   color: #f9a414;
