+++ src/Components/Layout/index.jsx
@@ -1,9 +1,11 @@
 import MemberView from 'Components/Views/memberView';
 import AdminView from 'Components/Views/adminView';
 import SuperAdminView from 'Components/Views/superAdminView';
+import LoginView from 'Components/Views/loginView';
+import { useSelector } from 'react-redux';
 
 const Layout = () => {
-  const role = sessionStorage.getItem('role');
+  const role = useSelector((state) => state.auth?.role);
 
   switch (role) {
     case 'SUPER_ADMIN':
@@ -12,8 +14,10 @@ const Layout = () => {
       return <AdminView />;
     case 'MEMBER':
       return <MemberView />;
+    case '':
+      return <LoginView />;
     default:
-      return <MemberView />;
+      return <LoginView />;
   }
 };
