+++ src/Components/Shared/Container/container.module.css
@@ -5,3 +5,11 @@
   align-items: center;
   justify-content: center;
 }
+
+.login {
+  display: flex;
+  flex-direction: column;
+  width: 100vw;
+  align-items: center;
+  justify-content: center;
+}
