+++ src/Components/Shared/Container/index.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import styles from './container.module.css';
 
-const Container = ({ children }) => <div className={styles.container}>{children}</div>;
+const Container = ({ children, isLogin }) => (
+  <div className={isLogin ? styles.login : styles.container}>{children}</div>
+);
 
 export default Container;
