--- src/Components/Views/trainerView.jsx
@@ -1,1 +1,0 @@
-///;
