+++ src/Routes/index.js
@@ -3,6 +3,7 @@ import { Switch, Route, BrowserRouter, Redirect } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import { tokenListener } from '../Helper/firebase';
 import { getAuth } from '../Redux/auth/thunks';
+
 import PrivateRoute from './privateRoute';
 
 const AdminRoutes = lazy(() => import('./admin'));
@@ -28,9 +29,9 @@ const Routes = () => {
     <BrowserRouter>
       <Suspense fallback={<div />}>
         <Switch>
-          <PrivateRoute path="/admin" role="ADMIN" component={AdminRoutes} />
-          <PrivateRoute path="/member" role="MEMBER" component={MemberRoutes} />
-          <PrivateRoute path="/super-admin" role="SUPER_ADMIN" component={SuperAdminRoutes} />
+          <PrivateRoute path="/admins" role="ADMIN" component={AdminRoutes} />
+          <PrivateRoute path="/members" role="MEMBER" component={MemberRoutes} />
+          <PrivateRoute path="/super-admins" role="SUPER_ADMIN" component={SuperAdminRoutes} />
           <Route path="/auth" component={AuthRoutes} />
           <Redirect to="/auth" />
         </Switch>
